+++ packages/runtime/package.json
@@ -29,15 +29,16 @@
     "build": "tsc"
   },
   "devDependencies": {
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "regenerator-runtime": "^0.13.9"
   },
   "sideEffects": false,
   "dependencies": {
     "@ice/jsx-runtime": "^0.1.0",
+    "@ice/types": "workspace:^1.0.0",
     "consola": "^2.15.3",
     "history": "^5.3.0",
     "react-router-dom": "^6.2.2"
