+++ packages/runtime/src/App.tsx
@@ -1,10 +1,10 @@
 import React, { useMemo } from 'react';
 import type { Action, Location } from 'history';
 import type { Navigator } from 'react-router-dom';
+import type { RouteWrapperConfig, AppRouterProps } from '@ice/types';
 import AppErrorBoundary from './AppErrorBoundary.js';
 import { useAppContext } from './AppContext.js';
 import { createRouteElements } from './routes.js';
-import type { RouteWrapperConfig, AppRouterProps } from './types.js';
 
 interface Props {
   action: Action;
