+++ packages/runtime/src/matchRoutes.ts
@@ -4,8 +4,8 @@
 import type { Location } from 'history';
 import type { RouteObject } from 'react-router-dom';
 import { matchRoutes as originMatchRoutes } from 'react-router-dom';
+import type { RouteItem, RouteMatch } from '@ice/types';
 import { matchRoutes as matchRoutesSingle } from './single-router.js';
-import type { RouteItem, RouteMatch } from './types.js';
 
 export default function matchRoutes(
   routes: RouteItem[],
