+++ packages/types/package.json
@@ -30,15 +30,17 @@
   "bugs": "https://github.com/ice-lab/ice-next/issues",
   "homepage": "https://next.ice.work",
   "devDependencies": {
-    "@swc/core": "1.2.210",
     "@ice/route-manifest": "^1.0.0",
-    "@ice/runtime": "^1.0.0",
+    "@swc/core": "1.2.210",
     "build-scripts": "^2.0.0-24",
     "esbuild": "^0.14.51",
     "eslint": "^8.14.0",
     "eslint-webpack-plugin": "3.1.1",
     "fork-ts-checker-webpack-plugin": "7.2.6",
+    "history": "^5.3.0",
     "react": "^18.0.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.2.2",
     "terser": "^5.12.1",
     "typescript": "^4.6.4",
     "unplugin": "^0.9.5",
